requirements.txt
einops
faster-coco-eval
graphviz
hydra-core
lightning
loguru
numpy
opencv-python
Pillow
pycocotools
requests
rich
torch
torchvision
wandb

# Docker環境用追加依存関係
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
onnx>=1.15.0
onnxruntime>=1.16.0
psutil>=5.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
jinja2>=3.1.0
tqdm>=4.66.0
